+++ PowerShell Scripts/Build Project.ps1
@@ -1,48 +1,65 @@
-# Clear the screen
-Clear-Host
+# Get arguments
+$MCGradleArg = $args[0]
 
-$MyPWSHAuthor = "Jonathing"
-$MyPWSHVersion = "0.4.1"
+if ($MCGradleArg -ne "FromHub")
+{
+    # Clear the screen
+    Clear-Host
 
-# Print script information
-$MyPWSHGreeting1 = "MCGradle Scripts (for Windows PowerShell and PowerShell Core)"
-$MyPWSHGreeting2 = "Version " + $MyPWSHVersion
-$MyPWSHGreeting3 = "Written and Maintained by " + $MyPWSHAuthor
-Write-Host $MyPWSHGreeting1
-Write-Host $MyPWSHGreeting2
-Write-Host $MyPWSHGreeting3
-Write-Host ""
+    $MCGradleAuthor = "Jonathing"
+    $MCGradleVersion = "0.5.0"
 
-# Go to root project directory
-Set-Location ..\..
+    # Print script information
+    $MCGradleGreeting1 = "MCGradle Scripts"
+    $MCGradleGreeting2 = "Version " + $MCGradleVersion
+    $MCGradleGreeting3 = "Written and Maintained by " + $MCGradleAuthor
+    Write-Host $MCGradleGreeting1
+    Write-Host $MCGradleGreeting2
+    Write-Host $MCGradleGreeting3
+    Write-Host ""
 
-# Check for update
-& '.\Scripts\PowerShell Scripts\internal\check_update.ps1' $MyPWSHVersion
+    # Go to root project directory
+    Set-Location ..\..
 
-# Get Forge mod name
-& '.\Scripts\PowerShell Scripts\internal\get_mod_name.ps1'
-$MyProjectName = Get-Content '.\Scripts\PowerShell Scripts\internal\MODNAME'
-Remove-Item '.\Scripts\PowerShell Scripts\internal\MODNAME'
+    # Check for update
+    & '.\Scripts\PowerShell Scripts\internal\check_update.ps1' $MCGradleVersion
+
+    # Get Forge mod name
+    & '.\Scripts\PowerShell Scripts\internal\get_mod_name.ps1'
+    $MCProjectName = Get-Content '.\Scripts\PowerShell Scripts\internal\MODNAME'
+    Remove-Item '.\Scripts\PowerShell Scripts\internal\MODNAME'
+}
 
 # Set the title of the Windows PowerShell or PowerShell Core console
-$MyPWSHTitle = $MyProjectName + ": Build Project"
-[System.Console]::Title = $MyPWSHTitle
+$MCGradleTitle = $MCProjectName + ": Build Project"
+[System.Console]::Title = $MCGradleTitle
 
 # Build the project
-$MyTaskMessage = "Building " + $MyProjectName + "..."
-Write-Host $MyTaskMessage
+$MCTaskMessage = "Building " + $MCProjectName + "..."
+Write-Host $MCTaskMessage
 Write-Host ""
 .\gradlew build --warning-mode none
 Write-Host ""
-$MyExitMessage = "Finished building " + $MyProjectName + "."
-Write-Host $MyExitMessage
-$MyExitMessage2 = "If the build was successful, the output should be located under build\libs"
-Write-Host $MyExitMessage2
+$MCExitMessage = "Finished building " + $MCProjectName + "."
+Write-Host $MCExitMessage
+$MCExitMessage2 = "If the build was successful, the output should be located under build\libs"
+Write-Host $MCExitMessage2
 
-# Return to scripts directory
-Set-Location '.\Scripts\PowerShell Scripts\'
+if ($MCGradleArg -ne "FromHub")
+{
+    # Return to scripts directory
+    Set-Location '.\Scripts\PowerShell Scripts\'
+}
 
 # END OF SCRIPT
 Pause
+
+if ($MCGradleArg -eq "FromHub")
+{
+    # Set the title of the Windows PowerShell or PowerShell Core console
+    $MCGradleTitle = $MCProjectName + ": MCGradle Scripts Hub"
+    [System.Console]::Title = $MCGradleTitle
+}
+
 Write-Host ""
 exit 0