+++ PowerShell Scripts/Set up IntelliJ.ps1
@@ -1,46 +1,63 @@
-# Clear the screen
-Clear-Host
+# Get arguments
+$MCGradleArg = $args[0]
 
-$MyPWSHAuthor = "Jonathing"
-$MyPWSHVersion = "0.4.1"
+if ($MCGradleArg -ne "FromHub")
+{
+    # Clear the screen
+    Clear-Host
 
-# Print script information
-$MyPWSHGreeting1 = "MCGradle Scripts (for Windows PowerShell and PowerShell Core)"
-$MyPWSHGreeting2 = "Version " + $MyPWSHVersion
-$MyPWSHGreeting3 = "Written and Maintained by " + $MyPWSHAuthor
-Write-Host $MyPWSHGreeting1
-Write-Host $MyPWSHGreeting2
-Write-Host $MyPWSHGreeting3
-Write-Host ""
+    $MCGradleAuthor = "Jonathing"
+    $MCGradleVersion = "0.5.0"
 
-# Go to root project directory
-Set-Location ..\..
+    # Print script information
+    $MCGradleGreeting1 = "MCGradle Scripts"
+    $MCGradleGreeting2 = "Version " + $MCGradleVersion
+    $MCGradleGreeting3 = "Written and Maintained by " + $MCGradleAuthor
+    Write-Host $MCGradleGreeting1
+    Write-Host $MCGradleGreeting2
+    Write-Host $MCGradleGreeting3
+    Write-Host ""
 
-# Check for update
-& '.\Scripts\PowerShell Scripts\internal\check_update.ps1' $MyPWSHVersion
+    # Go to root project directory
+    Set-Location ..\..
 
-# Get Forge mod name
-& '.\Scripts\PowerShell Scripts\internal\get_mod_name.ps1'
-$MyProjectName = Get-Content '.\Scripts\PowerShell Scripts\internal\MODNAME'
-Remove-Item '.\Scripts\PowerShell Scripts\internal\MODNAME'
+    # Check for update
+    & '.\Scripts\PowerShell Scripts\internal\check_update.ps1' $MCGradleVersion
+
+    # Get Forge mod name
+    & '.\Scripts\PowerShell Scripts\internal\get_mod_name.ps1'
+    $MCProjectName = Get-Content '.\Scripts\PowerShell Scripts\internal\MODNAME'
+    Remove-Item '.\Scripts\PowerShell Scripts\internal\MODNAME'
+}
 
 # Set the title of the Windows PowerShell console
-$MyPWSHTitle = $MyProjectName + ": IntelliJ IDEA Workspace"
-[System.Console]::Title = $MyPWSHTitle
+$MCGradleTitle = $MCProjectName + ": IntelliJ IDEA Workspace"
+[System.Console]::Title = $MCGradleTitle
 
 # Inform user that IntelliJ set up is done by IntelliJ IDEA
-$MyIntellijMessage1 = "The IntelliJ IDEA workspace for Forge is no longer set up through a command."
-$MyIntellijMessage2 = "To import the project to IntelliJ IDEA, simply open " + [char]0x0022 + "build.gradle" + [char]0x0022 + " as a project."
-$MyIntellijMessage3 = "Gradle will do the rest for you as it imports and indexes the project into IntelliJ."
-Write-Host $MyIntellijMessage1
-Write-Host $MyIntellijMessage2
-Write-Host $MyIntellijMessage3
+$MCIntellijMessage1 = "The IntelliJ IDEA workspace for Forge is no longer set up through a command."
+$MCIntellijMessage2 = "To import the project to IntelliJ IDEA, simply open " + [char]0x0022 + "build.gradle" + [char]0x0022 + " as a project."
+$MCIntellijMessage3 = "Gradle will do the rest for you as it imports and indexes the project into IntelliJ."
+Write-Host $MCIntellijMessage1
+Write-Host $MCIntellijMessage2
+Write-Host $MCIntellijMessage3
 Write-Host ""
 
-# Return to scripts directory
-Set-Location '.\Scripts\PowerShell Scripts\'
+if ($MCGradleArg -ne "FromHub")
+{
+    # Return to scripts directory
+    Set-Location '.\Scripts\PowerShell Scripts\'
+}
 
 # END OF SCRIPT
 Pause
+
+if ($MCGradleArg -eq "FromHub")
+{
+    # Set the title of the Windows PowerShell or PowerShell Core console
+    $MCGradleTitle = $MCProjectName + ": MCGradle Scripts Hub"
+    [System.Console]::Title = $MCGradleTitle
+}
+
 Write-Host ""
 exit 0