+++ bash Scripts/Build Project.sh
@@ -1,39 +1,50 @@
 #!/bin/bash
 
-# Clear the screen
-clear
+# Get arguments
+MCGradleArgs=$1
 
-MyBASHAuthor="Jonathing"
-MyBASHVersion="0.4.1"
+if [ "$MCGradleArgs" != "FromHub" ]
+then
+    # Clear the screen
+    clear
 
-# Print script information
-echo "MCGradle Scripts (for GNU bash)"
-echo "Version $MyBASHVersion"
-echo "Written and Maintained by $MyBASHAuthor"
-echo ""
+    MCGradleAuthor="Jonathing"
+    MCGradleVersion="0.5.0"
 
-# Check for update
-. ./internal/check_update.sh
+    # Print script information
+    echo "MCGradle Scripts"
+    echo "Version $MCGradleVersion"
+    echo "Written and Maintained by $MCGradleAuthor"
+    echo ""
 
-# Go to root project directory
-cd ../..
+    # Check for update
+    . ./internal/check_update.sh
 
-# Get Forge mod name
-MyProjectName=`grep 'displayName=' src/main/resources/META-INF/mods.toml -m 1`
-MyProjectName=${MyProjectName#*'"'}; MyProjectName=${MyProjectName%'"'*}
+    # Go to root project directory
+    cd ../..
+
+    # Get Forge mod name
+    MCProjectName=`grep 'displayName=' src/main/resources/META-INF/mods.toml -m 1`
+    MCProjectName=${MCProjectName#*'"'}; MCProjectName=${MCProjectName%'"'*}
+fi
 
 # Build the project
-echo "Building $MyProjectName..."
+echo "Building $MCProjectName..."
 echo ""
 ./gradlew build --warning-mode none
 echo ""
-echo "Finished building $MyProjectName."
+echo "Finished building $MCProjectName."
 echo "If the build was successful, the output should be located under build\libs"
 
-# Return to scripts directory
-cd Scripts/bash\ Scripts/
+if [ "$MCGradleArgs" != "FromHub" ]
+then
+    # Return to scripts directory
+    cd Scripts/bash\ Scripts/
+    read -s -n 1 -p "Press any key to exit MCGradle Scripts..."
+else
+    read -s -n 1 -p "Press any key to return to the MCGradle Scripts Hub..."
+fi
 
 # END OF SCRIPT
-read -s -n 1 -p "Press any key to exit MCGradle Scripts..."
 echo ""
-exit 0
+echo ""