+++ bash Scripts/Set up IntelliJ.sh
@@ -1,26 +1,32 @@
 #!/bin/bash
 
-# Clear the screen
-clear
+# Get arguments
+MCGradleArgs=$1
 
-MyBASHAuthor="Jonathing"
-MyBASHVersion="0.4.1"
+if [ "$MCGradleArgs" != "FromHub" ]
+then
+    # Clear the screen
+    clear
 
-# Print script information
-echo "MCGradle Scripts (for GNU bash)"
-echo "Version $MyBASHVersion"
-echo "Written and Maintained by $MyBASHAuthor"
-echo ""
+    MCGradleAuthor="Jonathing"
+    MCGradleVersion="0.5.0"
 
-# Check for update
-. ./internal/check_update.sh
+    # Print script information
+    echo "MCGradle Scripts"
+    echo "Version $MCGradleVersion"
+    echo "Written and Maintained by $MCGradleAuthor"
+    echo ""
 
-# Go to root project directory
-cd ../..
+    # Check for update
+    . ./internal/check_update.sh
 
-# Get Forge mod name
-MyProjectName=`grep 'displayName=' src/main/resources/META-INF/mods.toml -m 1`
-MyProjectName=${MyProjectName#*'"'}; MyProjectName=${MyProjectName%'"'*}
+    # Go to root project directory
+    cd ../..
+
+    # Get Forge mod name
+    MCProjectName=`grep 'displayName=' src/main/resources/META-INF/mods.toml -m 1`
+    MCProjectName=${MCProjectName#*'"'}; MCProjectName=${MCProjectName%'"'*}
+fi
 
 # Inform user that IntelliJ set up is done by IntelliJ IDEA
 echo "The IntelliJ IDEA workspace for Forge is no longer set up through a command."
@@ -28,10 +34,15 @@ echo "To import the project to IntelliJ IDEA, simply open the \"build.gradle\" f
 echo "Gradle will do the rest for you as it imports and indexes the project into IntelliJ."
 echo ""
 
-# Return to scripts directory
-cd Scripts/bash\ Scripts/
+if [ "$MCGradleArgs" != "FromHub" ]
+then
+    # Return to scripts directory
+    cd Scripts/bash\ Scripts/
+    read -s -n 1 -p "Press any key to exit MCGradle Scripts..."
+else
+    read -s -n 1 -p "Press any key to return to the MCGradle Scripts Hub..."
+fi
 
 # END OF SCRIPT
-read -s -n 1 -p "Press any key to exit MCGradle Scripts..."
 echo ""
-exit 0
+echo ""